fix(msc): prevent Positioned In/Out from becoming ARRIVAL
